add device option

+++ src/compressed_tensors/transform/factory/hadamard.py
@@ -60,7 +60,7 @@ class HadamardFactory(TransformFactory):
 
     def _create_weight(self, size: int, dtype: dtype, device: device) -> Parameter:
         data = deterministic_hadamard_matrix(size, dtype=dtype)
-        data = data.to(dtype=dtype, device=device)
+        data = data.to(device=device)
         return Parameter(data, requires_grad=self.scheme.requires_grad)
